pyproject.toml
[project]
name = "my-tracks"
version = "0.1.0"
description = "Django backend server for OwnTracks Android app"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14.0",
    "python-decouple>=3.8",
    "psycopg2-binary>=2.9.9",
    "gunicorn>=21.2.0",
    "channels>=4.0.0",
    "daphne>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "flake8>=6.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pyright>=1.1.0",
    "shellcheck-py>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mytracks", "tracker"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mytracks.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=tracker --cov-report=term-missing"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pyhamcrest>=2.1.0",
]
